All Views Added. Add and Edit for Vehicle Added
Things that work in this version - 
Booking View, Delete, Add and Update(Partially, Does not let you modify the passenger name for some reason)
Vehicle View, Delete, Add and update.
Routes View and Delete.

Things that do not work - 
Routes - Update and Add
Booking - Updating a Passenger name. 
Some of the cancel options for the views are not working correctly and will still display after they are used.