Checked and Tested Routes - They Work. Also implemented some Validation
Routes Edit and Add functions now work properly as intended. The last problem with Add/Edit is PassengerName in the bookings REST not updating on an Edit. Currently not sure whats causing this issue.
I have started to implement Angular and HTML validation. I just need to finished the Vehicles currently. Could add some more types of validation in the future including CSS validation to make things such as text box borders red if they are not valid.