Finished Validation. Started CSS and BootStrap.
Got the validation using Angular finished, I might just want to do a few HTML parts.
I know how to use CSS and BootStrap well for the website now and have implemented it on the booking parts of the website so far. Now I just need to clean up the style of the website in general and implement table and booking styles to the other parts of the website.